moving method arround
