+++ pages/hall-of-fame.vue
@@ -24,6 +24,7 @@
 
 <script>
   import axios from 'axios'
+  import steem from 'steem'
 
   import NavbarLoggedIn from '~/components/NavbarLoggedIn'
   import NavbarLoggedOut from '~/components/NavbarLoggedOut'
@@ -32,7 +33,7 @@
   import Contributor from '~/components/Contributor'
   import Delegator from '~/components/Delegator'
   import Curator from '~/components/Curator'
-
+ 
   import SteemConnect from '~/mixins/SteemConnect'
 
   export default {
@@ -97,7 +98,32 @@
       };
       let curators = await getCurators();
 
-      return { delegators, contributors, curators }; 
+      const getRsharesToSBDFactor = () => {
+        return new Promise((resolve, reject) => {
+          // get reward fund for posts
+          steem.api.getRewardFund('post', (err, fund) => {
+            if (err) reject(err);
+            else {
+              const rewardBalance = parseFloat(fund.reward_balance.replace(' STEEM', ''));
+              const recentClaims = parseInt(fund.recent_claims);
+
+              // get SBD price factor
+              steem.api.getCurrentMedianHistoryPrice((err, price) => {
+                if (err) reject(err);
+                else {
+                  const SBDPrice = parseFloat(price.base.replace(' SBD', ''));
+
+                  // calculate SBD value for each vote
+                  resolve(rewardBalance / recentClaims * SBDPrice);
+                  }
+                });
+              }
+            });
+          });
+        };
+      let rsharesToSBDFactor = await getRsharesToSBDFactor();
+
+      return { delegators, contributors, curators, rsharesToSBDFactor }; 
     },
     mounted() {
       // login via steemconnect (see: mixins/SteemConnect)
